refactor(bridge): simplify exact fee preview gating
Amp-Thread-ID: https://ampcode.com/threads/T-019d666d-8566-709d-b5df-9ab5bf028c0e
